test: criação de categoria como admin

+++ src/main/java/velaire/ecommmerce/business/dtos/ProductRequestDTO.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 public record ProductRequestDTO(
         String name,
         String description,
-        @NotNull @Positive BigDecimal price,
+        @NotNull @Positive double price,
         @NotNull @Min(0) Integer stockQuantity,
         @NotNull Long categoryId,
         @NotNull Long aromaId,
